Refactor version detection and unify address types

+++ source/memory.hpp
@@ -7,13 +7,13 @@ class Memory
 {
 public:
   template<typename T>
-  auto static ReadMemory(uintptr_t address) -> T
+  auto static ReadMemory(uint32_t address) -> T
   {
     return *reinterpret_cast<T*>(address);
   }
 
   template<typename T>
-  auto static inline WriteMemory(uintptr_t address, T new_value) -> void
+  auto static inline WriteMemory(uint32_t address, T new_value) -> void
   {
     *reinterpret_cast<T*>(address) = new_value;
   }
