+++ source/patches.cpp
@@ -21,9 +21,9 @@ Patches::IsGameRunning() -> bool
 }
 
 auto
-Patches::PatchHealthBar(int32_t ver_offset) -> bool
+Patches::PatchHealthBar(uint32_t ver_offset) -> bool
 {
-  uintptr_t address = 0;
+  uint32_t address = 0;
 
   if (!ver_offset)
     address = 0x00589395;
@@ -54,12 +54,12 @@ Patches::PatchHealthBar(int32_t ver_offset) -> bool
 
 struct OHKOArgs
 {
-  int32_t ver_offset;
+  uint32_t ver_offset;
   bool die_after_bike_fall;
 };
 
 auto
-Patches::StartOHKOThread(int32_t ver_offset, bool die_after_bike_fall) -> bool
+Patches::StartOHKOThread(uint32_t ver_offset, bool die_after_bike_fall) -> bool
 {
   auto* args = new OHKOArgs{ ver_offset, die_after_bike_fall };
   auto handle =
@@ -80,9 +80,9 @@ Patches::OHKOThread(LPVOID param) -> DWORD
   auto args = *reinterpret_cast<OHKOArgs*>(param);
   delete reinterpret_cast<OHKOArgs*>(param);
 
-  uintptr_t max_health_address = 0;
-  uintptr_t max_armour_address = 0;
-  uintptr_t player_pointer = 0;
+  uint32_t max_health_address = 0;
+  uint32_t max_armour_address = 0;
+  uint32_t player_pointer = 0;
 
   if (args.ver_offset == 0x75130) {
     max_health_address = 0x00C0BDC8;
@@ -95,22 +95,22 @@ Patches::OHKOThread(LPVOID param) -> DWORD
   }
 
   while (true) {
-    auto player = Memory::ReadMemory<uintptr_t>(player_pointer);
+    auto player = Memory::ReadMemory<uint32_t>(player_pointer);
 
     if (!player) {
       Sleep(20);
       continue;
     }
 
-    static constexpr uintptr_t k_health_offset = 0x540;
-    auto health_address = uintptr_t{ player + k_health_offset };
+    static constexpr uint32_t k_health_offset = 0x540;
+    auto health_address = uint32_t{ player + k_health_offset };
 
     if (args.die_after_bike_fall) {
-      static constexpr uintptr_t k_vehicle_model_id_offset = 0x22;
+      static constexpr uint32_t k_vehicle_model_id_offset = 0x22;
       static constexpr float k_crash_hp_delta = 0.15f;
-      static uintptr_t last_vehicle = 0;
+      static uint32_t last_vehicle = 0;
       static float last_health = 1.0001f;
-      auto current_vehicle = Memory::ReadMemory<uintptr_t>(0x00BA18FC);
+      auto current_vehicle = Memory::ReadMemory<uint32_t>(0x00BA18FC);
       auto current_health = Memory::ReadMemory<float>(health_address);
 
       if (last_vehicle != 0 && current_vehicle == 0) {
@@ -130,7 +130,7 @@ Patches::OHKOThread(LPVOID param) -> DWORD
       last_health = current_health;
     }
 
-    auto armour_address = uintptr_t{ player + 0x548 };
+    auto armour_address = uint32_t{ player + 0x548 };
 
     if (Memory::ReadMemory<float>(max_health_address) > 5.682f)
       *reinterpret_cast<float*>(max_health_address) = 5.682f;
